move windowDimensionsPropTypes
