+++ src/components/AnchorForCommentsOnly/BaseAnchorForCommentsOnly/index.js
@@ -3,10 +3,10 @@ import React from 'react';
 import {Pressable, StyleSheet} from 'react-native';
 import lodashGet from 'lodash/get';
 import Text from '../../Text';
-import {propTypes, defaultProps} from '../anchorForCommentsOnlyPropTypes';
+import * as anchorForCommentsOnlyPropTypes from '../anchorForCommentsOnlyPropTypes';
 import PressableWithSecondaryInteraction from '../../PressableWithSecondaryInteraction';
-import {showContextMenu} from '../../../pages/home/report/ContextMenu/ReportActionContextMenu';
-import {CONTEXT_MENU_TYPES} from '../../../pages/home/report/ContextMenu/ContextMenuActions';
+import * as ReportActionContextMenu from '../../../pages/home/report/ContextMenu/ReportActionContextMenu';
+import * as ContextMenuActions from '../../../pages/home/report/ContextMenu/ContextMenuActions';
 import AttachmentView from '../../AttachmentView';
 import fileDownload from '../../../libs/fileDownload';
 
@@ -15,7 +15,7 @@ import fileDownload from '../../../libs/fileDownload';
  */
 const BaseAnchorForCommentsOnly = (props) => {
     let linkRef;
-    const rest = _.omit(props, _.keys(propTypes));
+    const rest = _.omit(props, _.keys(anchorForCommentsOnlyPropTypes.propTypes));
     return (
         props.isAttachment
             ? (
@@ -34,8 +34,8 @@ const BaseAnchorForCommentsOnly = (props) => {
                 <PressableWithSecondaryInteraction
                     onSecondaryInteraction={
                             (event) => {
-                                showContextMenu(
-                                    CONTEXT_MENU_TYPES.LINK,
+                                ReportActionContextMenu.showContextMenu(
+                                    ContextMenuActions.CONTEXT_MENU_TYPES.LINK,
                                     event,
                                     props.href,
                                     lodashGet(linkRef, 'current'),
@@ -63,8 +63,8 @@ const BaseAnchorForCommentsOnly = (props) => {
     );
 };
 
-BaseAnchorForCommentsOnly.propTypes = propTypes;
-BaseAnchorForCommentsOnly.defaultProps = defaultProps;
+BaseAnchorForCommentsOnly.propTypes = anchorForCommentsOnlyPropTypes.propTypes;
+BaseAnchorForCommentsOnly.defaultProps = anchorForCommentsOnlyPropTypes.defaultProps;
 BaseAnchorForCommentsOnly.displayName = 'BaseAnchorForCommentsOnly';
 
 export default BaseAnchorForCommentsOnly;
