+++ src/components/AnchorForCommentsOnly/BaseAnchorForCommentsOnly/index.native.js
@@ -2,12 +2,12 @@ import _ from 'underscore';
 import React from 'react';
 import lodashGet from 'lodash/get';
 import {Linking, StyleSheet, Pressable} from 'react-native';
-import {propTypes, defaultProps} from '../anchorForCommentsOnlyPropTypes';
+import * as anchorForCommentsOnlyPropTypes from '../anchorForCommentsOnlyPropTypes';
 import fileDownload from '../../../libs/fileDownload';
 import Text from '../../Text';
 import PressableWithSecondaryInteraction from '../../PressableWithSecondaryInteraction';
-import {showContextMenu} from '../../../pages/home/report/ContextMenu/ReportActionContextMenu';
-import {CONTEXT_MENU_TYPES} from '../../../pages/home/report/ContextMenu/ContextMenuActions';
+import * as ReportActionContextMenu from '../../../pages/home/report/ContextMenu/ReportActionContextMenu';
+import * as ContextMenuActions from '../../../pages/home/report/ContextMenu/ContextMenuActions';
 import AttachmentView from '../../AttachmentView';
 import styles from '../../../styles/styles';
 
@@ -16,7 +16,7 @@ import styles from '../../../styles/styles';
  */
 const BaseAnchorForCommentsOnly = (props) => {
     let linkRef;
-    const rest = _.omit(props, _.keys(propTypes));
+    const rest = _.omit(props, _.keys(anchorForCommentsOnlyPropTypes.propTypes));
     return (
         props.isAttachment
             ? (
@@ -37,8 +37,8 @@ const BaseAnchorForCommentsOnly = (props) => {
                 <PressableWithSecondaryInteraction
                     onSecondaryInteraction={
                 (event) => {
-                    showContextMenu(
-                        CONTEXT_MENU_TYPES.LINK,
+                    ReportActionContextMenu.showContextMenu(
+                        ContextMenuActions.CONTEXT_MENU_TYPES.LINK,
                         event,
                         props.href,
                         lodashGet(linkRef, 'current'),
@@ -60,8 +60,8 @@ const BaseAnchorForCommentsOnly = (props) => {
     );
 };
 
-BaseAnchorForCommentsOnly.propTypes = propTypes;
-BaseAnchorForCommentsOnly.defaultProps = defaultProps;
+BaseAnchorForCommentsOnly.propTypes = anchorForCommentsOnlyPropTypes.propTypes;
+BaseAnchorForCommentsOnly.defaultProps = anchorForCommentsOnlyPropTypes.defaultProps;
 BaseAnchorForCommentsOnly.displayName = 'BaseAnchorForCommentsOnly';
 
 export default BaseAnchorForCommentsOnly;
