+++ src/components/AnchorForCommentsOnly/index.js
@@ -1,10 +1,7 @@
 import _ from 'underscore';
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-    propTypes as anchorForCommentsOnlyPropTypes,
-    defaultProps as anchorForCommentsOnlyDefaultProps,
-} from './anchorForCommentsOnlyPropTypes';
+import * as anchorForCommentsOnlyPropTypes from './anchorForCommentsOnlyPropTypes';
 import BaseAnchorForCommentsOnly from './BaseAnchorForCommentsOnly';
 import addEncryptedAuthTokenToURL from '../../libs/addEncryptedAuthTokenToURL';
 
@@ -12,12 +9,12 @@ const propTypes = {
     /** Do we need an auth token to view this link or download the remote resource? */
     isAuthTokenRequired: PropTypes.bool,
 
-    ...anchorForCommentsOnlyPropTypes,
+    ...anchorForCommentsOnlyPropTypes.propTypes,
 };
 
 const defaultProps = {
     isAuthTokenRequired: false,
-    ...anchorForCommentsOnlyDefaultProps,
+    ...anchorForCommentsOnlyPropTypes.defaultProps,
 };
 
 /*
