+++ src/components/CheckboxWithTooltip/checkboxWithTooltipPropTypes.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import {windowDimensionsPropTypes} from '../withWindowDimensions';
+import windowDimensionsPropTypes from '../withWindowDimensions/windowDimensionsPropTypes';
 import CONST from '../../CONST';
 import stylePropTypes from '../../styles/stylePropTypes';
