+++ src/components/withWindowDimensions/windowDimensionsPropTypes.js
@@ -1,0 +1,15 @@
+import PropTypes from 'prop-types';
+
+export default {
+    // Width of the window
+    windowWidth: PropTypes.number.isRequired,
+
+    // Height of the window
+    windowHeight: PropTypes.number.isRequired,
+
+    // Is the window width narrow, like on a mobile device?
+    isSmallScreenWidth: PropTypes.bool.isRequired,
+
+    // Is the window width narrow, like on a tablet device?
+    isMediumScreenWidth: PropTypes.bool.isRequired,
+};
